adds support for ordered fieldlist in result set

+++ src/main/java/oakgit/engine/query/analyzer/UpdateAnalyzer.java
@@ -60,7 +60,7 @@ public class UpdateAnalyzer implements QueryAnalyzer {
                             };
                             break;
                         case "DSIZE + ?":
-                            Long sizeAddendum = (Long) placeholderData.get(placeHolderIndex);
+                            Integer sizeAddendum = (Integer) placeholderData.get(placeHolderIndex);
                             placeHolderIndex++;
                             updateValue = (Function<DocumentEntry, Long>) entry ->
                                 (entry.getDSize() != null ? entry.getDSize() : 0L) +
